Update StandardController.php
When cancelling an order, the customer can use the back button in the browser, and finalize the payment anyways. This will set the order in the "processing" state, but will leave the product items in the cancelled state, rendering the order useless. Store owners cannot cancel or finalize the order through the backend, and MySQL work is required to cancel the order completely.

This way, the order is left in "pending" state if the customer doesn't pay - and in "processing" state with "working" product items if the customer uses the back button and pays anyways.